feat: make a base protocol for rattler-build and pixi
