+++ crates/pixi_build_frontend/Cargo.toml
@@ -10,6 +10,7 @@ repository.workspace = true
 version = "0.1.0"
 
 [dependencies]
+async-trait = { workspace = true }
 dashmap = { workspace = true }
 futures = { workspace = true }
 itertools = { workspace = true }
